requirements.txt
django==4.2.13
djangorestframework==3.15.1
django-cors-headers==4.3.1
requests==2.31.0
beautifulsoup4==4.12.3
selenium==4.18.1
webdriver-manager==4.0.1
anthropic==0.28.0
chromadb==0.5.0
sentence-transformers==3.0.1
python-dotenv==1.0.1
Pillow==10.3.0
celery==5.3.6
redis==5.0.4
django-redis==5.4.0
mysqlclient==2.2.4
